Small refactoring of the Lexer/Parser: - LexerParser is now a subfolder of `sequences/` - Separation between ListParser and CalculateTree - Added docstrings for almost every function (subject to rewriting   anyway)
