+++ sardine/sequences/LexerParser/Qualifiers.py
@@ -1,0 +1,99 @@
+qualifiers = {
+    "dim": [0, 3, 6, 12],
+    "dim9": [0, 3, 6, 9, 14],
+    "hdim7": [0, 3, 6, 10],
+    "hdim9": [0, 3, 6, 10, 14],
+    "hdimb9": [0, 3, 6, 10, 13],
+    "dim7": [0, 3, 6, 9],
+    "7dim5": [0, 4, 6, 10],
+    "aug": [0, 4, 8, 12],
+    "augMaj7": [0, 4, 8, 11],
+    "aug7": [0, 4, 8, 10],
+    "aug9": [0, 4, 10, 14],
+    "maj": [0, 4, 7, 12],
+    "maj7": [0, 4, 7, 11],
+    "maj9": [0, 4, 11, 14],
+    "minmaj7": [0, 3, 7, 11],
+    "5": [0, 7, 12],
+    "6": [0, 4, 7, 9],
+    "7": [0, 4, 7, 10],
+    "9": [0, 4, 10, 14],
+    "b9": [0, 4, 10, 13],
+    "mM9": [0, 3, 11, 14],
+    "min": [0, 3, 7, 12],
+    "min7": [0, 3, 7, 10],
+    "min9": [0, 3, 10, 14],
+    "sus4": [0, 5, 7, 12],
+    "sus2": [0, 2, 7, 12],
+    "b5": [0, 4, 6, 12],
+    "mb5": [0, 3, 6, 12],
+    # Scales begin here
+    # Based on a very partial list found here:
+    # https://en.wikipedia.org/wiki/List_of_musical_scales_and_modes
+    "major": [0, 2, 4, 5, 7, 9, 11],
+    "minor": [0, 2, 3, 5, 7, 8, 10],
+    "hminor": [0, 2, 3, 5, 7, 8, 11],
+    "^minor": [0, 2, 3, 5, 7, 9, 11],  # doesn't work
+    "vminor": [0, 2, 3, 5, 7, 8, 10],
+    "penta": [0, 2, 4, 7, 9],
+    "acoustic": [0, 2, 4, 6, 7, 9, 10],
+    "aeolian": [0, 2, 3, 5, 7, 8, 10],
+    "algerian": [0, 2, 3, 6, 7, 9, 11, 12, 14, 15, 17],
+    "superlocrian": [0, 1, 3, 4, 6, 8, 10],
+    "augmented": [0, 3, 4, 7, 8, 11],
+    "bebop": [0, 2, 4, 5, 7, 9, 10, 11],
+    "blues": [0, 3, 5, 6, 7, 10],
+    "chromatic": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
+    "dorian": [0, 2, 3, 5, 7, 9, 10],
+    "double-harmonic": [0, 1, 4, 5, 8, 11],
+    "enigmatic": [0, 1, 4, 6, 8, 10, 11],
+    "flamenco": [0, 1, 4, 5, 7, 8, 11],
+    "gypsy": [0, 2, 3, 6, 7, 8, 10],
+    "halfdim": [0, 2, 3, 5, 6, 8, 10],
+    "harm-major": [0, 2, 4, 5, 7, 8, 11],
+    "harm-minor": [0, 2, 3, 5, 7, 8, 11],
+    "hirajoshi": [0, 4, 6, 7, 11],
+    "hungarian-minor": [0, 2, 3, 6, 7, 8, 11],
+    "hungarian-major": [0, 3, 4, 6, 7, 9, 10],
+    "in": [0, 1, 5, 7, 8],
+    "insen": [0, 1, 5, 7, 10],
+    "ionian": [0, 2, 4, 5, 7, 9, 11],
+    "istrian": [0, 1, 3, 4, 6, 7],
+    "iwato": [0, 1, 5, 6, 10],
+    "locrian": [0, 1, 3, 5, 6, 8, 10],
+    "lydian-augmented": [0, 2, 4, 6, 8, 9, 11],
+    "lydian": [0, 2, 4, 5, 7, 8, 9, 11],
+    "major-locrian": [0, 2, 4, 5, 6, 8, 10],
+    "major-penta": [0, 2, 4, 7, 9],
+    "melodic-minor-ascending": [0, 2, 3, 5, 7, 9, 11],
+    "melodic-minor-descending": [0, 2, 3, 5, 7, 8, 10],
+    "minor-penta": [0, 3, 5, 7, 10],
+    "mixolydian": [0, 2, 4, 5, 7, 9, 10],
+    "neapolitan": [0, 1, 3, 5, 7, 8, 11],
+    "octatonic": [0, 2, 3, 5, 6, 8, 9, 11],
+    "octatonic2": [0, 1, 3, 4, 6, 7, 9, 10],
+    "persian": [0, 1, 4, 5, 6, 8, 11],
+    "phrygian": [0, 1, 4, 5, 7, 8, 10],
+    "prometheus": [0, 2, 4, 6, 9, 10],
+    "harmonics": [0, 3, 4, 5, 7, 9],
+    "tritone": [0, 1, 4, 6, 7, 10],
+    "two-semitone": [0, 1, 2, 6, 7, 8],
+    "ukrainian": [0, 2, 3, 6, 7, 9, 10],
+    "whole": [0, 2, 4, 6, 8, 10],
+    "yo": [0, 3, 5, 7, 10],
+    "symetrical": [0, 1, 2, 6, 7, 10],
+    "symetrical2": [0, 2, 3, 6, 8, 10],
+    "messiaen1": [0, 2, 4, 6, 8, 10],
+    "messiaen2": [0, 1, 3, 4, 6, 7, 9, 10],
+    "messiaen3": [0, 2, 3, 4, 6, 7, 8, 10, 11],
+    "messiaen4": [0, 1, 2, 4, 6, 7, 8, 11],
+    "messiaen5": [0, 1, 5, 6, 7, 11],
+    "messiaen6": [0, 2, 4, 5, 6, 8],
+    "messiaen7": [0, 1, 2, 3, 5, 6, 7, 8, 9, 11],
+    # Structures (other musical objects)
+    "fourths": [0, 4, 10, 15, 20],
+    "fifths": [0, 7, 14, 21, 28],
+    "sixths": [0, 9, 17, 26, 35],
+    "thirds": [0, 4, 8, 12],
+    "octaves": [0, 12, 24, 36, 48],
+}
