+++ sardine/sequences/LexerParser/TreeCalc.py
@@ -1,123 +1,10 @@
-from lark import Lark, Transformer, v_args, Tree
-from itertools import cycle, islice, count, chain
-from pathlib import Path
+from lark import Transformer, v_args
+from itertools import cycle
+from .Qualifiers import qualifiers
 import random
 
-__all__ = ("ListParser", "Pnote", "Pname", "Pnum")
 
-
-class ParserError(Exception):
-    pass
-
-
-qualifiers = {
-    "dim": [0, 3, 6, 12],
-    "dim9": [0, 3, 6, 9, 14],
-    "hdim7": [0, 3, 6, 10],
-    "hdim9": [0, 3, 6, 10, 14],
-    "hdimb9": [0, 3, 6, 10, 13],
-    "dim7": [0, 3, 6, 9],
-    "7dim5": [0, 4, 6, 10],
-    "aug": [0, 4, 8, 12],
-    "augMaj7": [0, 4, 8, 11],
-    "aug7": [0, 4, 8, 10],
-    "aug9": [0, 4, 10, 14],
-    "maj": [0, 4, 7, 12],
-    "maj7": [0, 4, 7, 11],
-    "maj9": [0, 4, 11, 14],
-    "minmaj7": [0, 3, 7, 11],
-    "5": [0, 7, 12],
-    "6": [0, 4, 7, 9],
-    "7": [0, 4, 7, 10],
-    "9": [0, 4, 10, 14],
-    "b9": [0, 4, 10, 13],
-    "mM9": [0, 3, 11, 14],
-    "min": [0, 3, 7, 12],
-    "min7": [0, 3, 7, 10],
-    "min9": [0, 3, 10, 14],
-    "sus4": [0, 5, 7, 12],
-    "sus2": [0, 2, 7, 12],
-    "b5": [0, 4, 6, 12],
-    "mb5": [0, 3, 6, 12],
-    # Scales begin here
-    # Based on a very partial list found here:
-    # https://en.wikipedia.org/wiki/List_of_musical_scales_and_modes
-    "major": [0, 2, 4, 5, 7, 9, 11],
-    "minor": [0, 2, 3, 5, 7, 8, 10],
-    "hminor": [0, 2, 3, 5, 7, 8, 11],
-    "^minor": [0, 2, 3, 5, 7, 9, 11],  # doesn't work
-    "vminor": [0, 2, 3, 5, 7, 8, 10],
-    "penta": [0, 2, 4, 7, 9],
-    "acoustic": [0, 2, 4, 6, 7, 9, 10],
-    "aeolian": [0, 2, 3, 5, 7, 8, 10],
-    "algerian": [0, 2, 3, 6, 7, 9, 11, 12, 14, 15, 17],
-    "superlocrian": [0, 1, 3, 4, 6, 8, 10],
-    "augmented": [0, 3, 4, 7, 8, 11],
-    "bebop": [0, 2, 4, 5, 7, 9, 10, 11],
-    "blues": [0, 3, 5, 6, 7, 10],
-    "chromatic": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
-    "dorian": [0, 2, 3, 5, 7, 9, 10],
-    "double-harmonic": [0, 1, 4, 5, 8, 11],
-    "enigmatic": [0, 1, 4, 6, 8, 10, 11],
-    "flamenco": [0, 1, 4, 5, 7, 8, 11],
-    "gypsy": [0, 2, 3, 6, 7, 8, 10],
-    "halfdim": [0, 2, 3, 5, 6, 8, 10],
-    "harm-major": [0, 2, 4, 5, 7, 8, 11],
-    "harm-minor": [0, 2, 3, 5, 7, 8, 11],
-    "hirajoshi": [0, 4, 6, 7, 11],
-    "hungarian-minor": [0, 2, 3, 6, 7, 8, 11],
-    "hungarian-major": [0, 3, 4, 6, 7, 9, 10],
-    "in": [0, 1, 5, 7, 8],
-    "insen": [0, 1, 5, 7, 10],
-    "ionian": [0, 2, 4, 5, 7, 9, 11],
-    "istrian": [0, 1, 3, 4, 6, 7],
-    "iwato": [0, 1, 5, 6, 10],
-    "locrian": [0, 1, 3, 5, 6, 8, 10],
-    "lydian-augmented": [0, 2, 4, 6, 8, 9, 11],
-    "lydian": [0, 2, 4, 5, 7, 8, 9, 11],
-    "major-locrian": [0, 2, 4, 5, 6, 8, 10],
-    "major-penta": [0, 2, 4, 7, 9],
-    "melodic-minor-ascending": [0, 2, 3, 5, 7, 9, 11],
-    "melodic-minor-descending": [0, 2, 3, 5, 7, 8, 10],
-    "minor-penta": [0, 3, 5, 7, 10],
-    "mixolydian": [0, 2, 4, 5, 7, 9, 10],
-    "neapolitan": [0, 1, 3, 5, 7, 8, 11],
-    "octatonic": [0, 2, 3, 5, 6, 8, 9, 11],
-    "octatonic2": [0, 1, 3, 4, 6, 7, 9, 10],
-    "persian": [0, 1, 4, 5, 6, 8, 11],
-    "phrygian": [0, 1, 4, 5, 7, 8, 10],
-    "prometheus": [0, 2, 4, 6, 9, 10],
-    "harmonics": [0, 3, 4, 5, 7, 9],
-    "tritone": [0, 1, 4, 6, 7, 10],
-    "two-semitone": [0, 1, 2, 6, 7, 8],
-    "ukrainian": [0, 2, 3, 6, 7, 9, 10],
-    "whole": [0, 2, 4, 6, 8, 10],
-    "yo": [0, 3, 5, 7, 10],
-    "symetrical": [0, 1, 2, 6, 7, 10],
-    "symetrical2": [0, 2, 3, 6, 8, 10],
-    "messiaen1": [0, 2, 4, 6, 8, 10],
-    "messiaen2": [0, 1, 3, 4, 6, 7, 9, 10],
-    "messiaen3": [0, 2, 3, 4, 6, 7, 8, 10, 11],
-    "messiaen4": [0, 1, 2, 4, 6, 7, 8, 11],
-    "messiaen5": [0, 1, 5, 6, 7, 11],
-    "messiaen6": [0, 2, 4, 5, 6, 8],
-    "messiaen7": [0, 1, 2, 3, 5, 6, 7, 8, 9, 11],
-    # Structures (other musical objects)
-    "fourths": [0, 4, 10, 15, 20],
-    "fifths": [0, 7, 14, 21, 28],
-    "sixths": [0, 9, 17, 26, 35],
-    "thirds": [0, 4, 8, 12],
-    "octaves": [0, 12, 24, 36, 48],
-}
-
-
-def floating_point_range(start, end, step):
-    assert step != 0
-    sample_count = int(abs(end - start) / step)
-    return islice(count(start, step), sample_count)
-
-
-@v_args(inline=True)  # Affects the signatures of the methods
+@v_args(inline=True)
 class CalculateTree(Transformer):
     number = float
 
@@ -398,7 +285,8 @@ class CalculateTree(Transformer):
             collection (list): A list generated through a qualifier
 
         Returns:
-            list: palindromed list of integers from qualifier's based collection
+            list: palindromed list of integers from qualifier's based
+            collection
         """
         collection.reverse()
         return collection + list(reversed(collection))
@@ -548,7 +436,8 @@ class CalculateTree(Transformer):
             collection (list): A list of integers
 
         Returns:
-            list: A list of integers with the second and fourth note dropped an octave.
+            list: A list of integers with the second and fourth note dropped
+            an octave.
         """
         collection[1] = collection[1] - 12
         collection[3] = collection[3] - 12
@@ -749,157 +638,38 @@ class CalculateTree(Transformer):
         return a + b
 
     def sub_name(self, a, b):
+        """Substraction of a name by a name. Every letter present
+        in name 'b' will be substracted from name 'a'.
+
+        Args:
+            a (str): A name
+            b (str): A name
+
+        Returns:
+            str: A Substracted name.
+        """
         return a.replace(b, "")
 
     def choice_name(self, a, b):
+        """Choose 50%/50% between name 'a' and name 'b'.
+
+        Args:
+            a (str): A name
+            b (str): A name
+
+        Returns:
+            str: The name chosen by a chance operation
+        """
         return random.choice([a, b])
 
     def repeat_name(self, name, value):
+        """Repeats a name 'value' times.
+
+        Args:
+            name (str): A name given as a string
+            value (_type_): Number of repetitions
+
+        Returns:
+            list[str]: A list composed of 'value' times the 'name'.
+        """
         return [name] * int(value)
-
-
-grammar_path = Path(__file__).parent
-grammars = {
-    "number": grammar_path / "grammars/number.lark",
-    "name": grammar_path / "grammars/name.lark",
-    "note": grammar_path / "grammars/note.lark",
-}
-
-parsers = {
-    "number": {
-        "raw": Lark.open(
-            grammars["number"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-        ),
-        "full": Lark.open(
-            grammars["number"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-            transformer=CalculateTree(),
-        ),
-    },
-    "name": {
-        "raw": Lark.open(
-            grammars["name"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-        ),
-        "full": Lark.open(
-            grammars["name"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-            transformer=CalculateTree(),
-        ),
-    },
-    "note": {
-        "raw": Lark.open(
-            grammars["note"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-        ),
-        "full": Lark.open(
-            grammars["note"],
-            rel_to=__file__,
-            parser="lalr",
-            start="start",
-            cache=True,
-            lexer="contextual",
-            transformer=CalculateTree(),
-        ),
-    },
-}
-
-
-class ListParser:
-    def __init__(self, parser_type: str = "number"):
-        """
-        Initialise two different parsers:
-        - result_parser: the parser used to generate patterns
-        - printing_parser: the parser used for debugging purposes
-        """
-        try:
-            self._result_parser = parsers[parser_type]["full"]
-            self._printing_parser = parsers[parser_type]["raw"]
-        except KeyError:
-            ParserError(f"Invalid Parser grammar, {parser_type} is not a grammar.")
-
-    def _flatten_result(self, pat):
-        """Flatten a nested pattern result list. Probably not optimised."""
-        if len(pat) == 0:
-            return pat
-        if isinstance(pat[0], list):
-            return self._flatten_result(pat[0]) + self._flatten_result(pat[1:])
-        return pat[:1] + self._flatten_result(pat[1:])
-
-    def pretty_print(self, expression: str):
-        """Pretty print an expression from parser"""
-        print(f"EXPR: {expression}")
-        print(Tree.pretty(self._printing_parser.parse(expression)))
-        print(f"RESULT: {self._result_parser.parse(expression)}")
-
-    def print_tree_only(self, expression: str):
-        """Print only tree for debugging purposes"""
-        print(Tree.pretty(self._printing_parser.parse(expression)))
-
-    def _parse_token(self, string: str):
-        """Parse a single token and return the result for usage"""
-        return self._result_parser.parse(string)
-
-    def parse(self, pattern: str):
-        """Parse a whole pattern and return a flattened list"""
-        final_pattern = []
-        for token in pattern.split():
-            try:
-                final_pattern.append(self._parse_token(token))
-            except Exception as e:
-                raise ParserError(f"Non valid token: {token}") from e
-        return self._flatten_result(final_pattern)
-
-    def _parse_debug(self, pattern: str):
-        """Parse a whole pattern in debug mode"""
-        final_pattern = []
-        for token in pattern.split():
-            try:
-                self.pretty_print(expression=token)
-            except Exception as e:
-                import traceback
-
-                print(f"Error: {e}: {traceback.format_exc()}")
-                continue
-
-
-# Useful utilities
-
-
-def Pname(pattern: str, i: int = 0):
-    parser = ListParser(parser_type="name")
-    pattern = parser.parse(pattern)
-    return pattern[i % len(pattern)]
-
-
-def Pnote(pattern: str, i: int = 0):
-    parser = ListParser(parser_type="note")
-    pattern = parser.parse(pattern)
-    return pattern[i % len(pattern)]
-
-
-def Pnum(pattern: str, i: int = 0):
-    parser = ListParser(parser_type="number")
-    pattern = parser.parse(pattern)
-    return pattern[i % len(pattern)]
