+++ sardine/sequences/__init__.py
@@ -1,3 +1,3 @@
 from .Sequence import *
 from .Chance import *
-from .ListParser import *
+from .LexerParser.ListParser import *
